+++ next/src/comman/utils/today.ts
@@ -1,3 +1,5 @@
 import dayjs from "dayjs";
 
 export const today = dayjs();
+
+export const todayFormat = today.format("YYYY-MM");
