+++ next/src/store/atoms/comman.ts
@@ -1,4 +1,3 @@
 import { atom } from "jotai";
 
 export const is_homesplash_atom = atom<boolean>(true);
-export const is_loading_atom = atom<boolean>(false);
